requirements.txt
gunicorn
Werkzeug==0.9.4
Flask==0.10.1
flask_sqlalchemy
flask-login
flask-script
flask-fillin
psycopg2
Flask-SocketIO
Jinja2==2.7.2
MarkupSafe==0.18
greenlet==0.4.2
gevent==1.0
gevent-socketio==0.3.6
gevent-websocket==0.9.2
itsdangerous==0.23
ujson==1.33
python-bitcoinrpc
websocket-client
pandas
requests
pycrypto
bunch
py-bcrypt
hashlib
Crypto
beautifulsoup4
nose
coverage
Flask-Security
Flask-Social
http://github.com/pythonforfacebook/facebook-sdk/tarball/master
python-twitter
oauth2client
google-api-python-client
